requirements.txt
# Dépendances principales
flask==3.1.0
faiss-cpu==1.10.0
sentence-transformers==3.4.1
pydantic==2.10.6
numpy==1.26.4
python-dotenv==1.0.1
langchain==0.2.12
langchain-text-splitters==0.2.2
langchain-core>=0.2.27,<0.3.0  # Updated to resolve dependency conflicts

# Traitement des documents
pytesseract==0.3.13
PyMuPDF==1.25.3
python-docx==1.1.2
pandas==2.2.3
python-pptx==1.0.2
Pillow==11.1.0
opencv-python-headless==4.11.0.86

# Azure OpenAI
azure-ai-inference==1.0.0b9
azure-identity==1.20.0
azure-core==1.32.0

# Utilitaires
tqdm==4.66.4
rich==13.7.1
tabulate==0.9.0

# Sécurité et authentification
werkzeug>=3.1.0  # Updated to resolve dependency conflicts
itsdangerous==2.2.0
bcrypt==4.1.2

# Déploiement
gunicorn==22.0.0

# Azure Storage Blob
azure-storage-blob
pytest